Fix bug in final review when url is marked as new agency
