+++ tests/automated/integration/db/client/test_delete_url_updated_at.py
@@ -1,0 +1,24 @@
+from src.db.dtos.url_info import URLInfo
+from tests.helpers.db_data_creator import DBDataCreator
+
+
+def test_delete_url_updated_at(db_data_creator: DBDataCreator):
+    batch_id = db_data_creator.batch()
+    url_id = db_data_creator.urls(batch_id=batch_id, url_count=1).url_mappings[0].url_id
+
+    db_client = db_data_creator.db_client
+    url_info = db_client.get_urls_by_batch(batch_id=batch_id, page=1)[0]
+
+    old_updated_at = url_info.updated_at
+
+
+    db_client.update_url(
+        url_info=URLInfo(
+            id=url_id,
+            url="dg",
+            collector_metadata={"test_metadata": "test_metadata"},
+        )
+    )
+
+    url = db_client.get_urls_by_batch(batch_id=batch_id, page=1)[0]
+    assert url.updated_at > old_updated_at
