+++ tests/automated/integration/db/client/test_insert_logs.py
@@ -1,0 +1,26 @@
+import pytest
+
+from src.db.dtos.log_info import LogInfo
+from tests.helpers.db_data_creator import DBDataCreator
+
+
+@pytest.mark.asyncio
+async def test_insert_logs(db_data_creator: DBDataCreator):
+    batch_id_1 = db_data_creator.batch()
+    batch_id_2 = db_data_creator.batch()
+
+    adb_client = db_data_creator.adb_client
+    db_client = db_data_creator.db_client
+    db_client.insert_logs(
+        log_infos=[
+            LogInfo(log="test log", batch_id=batch_id_1),
+            LogInfo(log="test log", batch_id=batch_id_1),
+            LogInfo(log="test log", batch_id=batch_id_2),
+        ]
+    )
+
+    logs = await adb_client.get_logs_by_batch_id(batch_id_1)
+    assert len(logs) == 2
+
+    logs = await adb_client.get_logs_by_batch_id(batch_id_2)
+    assert len(logs) == 1
