+++ tests/automated/integration/db/client/test_insert_urls.py
@@ -1,0 +1,48 @@
+import pytest
+
+from src.core.enums import BatchStatus
+from src.db.dtos.batch_info import BatchInfo
+from src.db.dtos.url_info import URLInfo
+
+
+@pytest.mark.asyncio
+async def test_insert_urls(
+        db_client_test,
+        adb_client_test
+):
+    # Insert batch
+    batch_info = BatchInfo(
+        strategy="ckan",
+        status=BatchStatus.IN_PROCESS,
+        parameters={},
+        user_id=1
+    )
+    batch_id = await adb_client_test.insert_batch(batch_info)
+
+    urls = [
+        URLInfo(
+            url="https://example.com/1",
+            collector_metadata={"name": "example_1"},
+        ),
+        URLInfo(
+            url="https://example.com/2",
+        ),
+        # Duplicate
+        URLInfo(
+            url="https://example.com/1",
+            collector_metadata={"name": "example_duplicate"},
+        )
+    ]
+    insert_urls_info = await adb_client_test.insert_urls(
+        url_infos=urls,
+        batch_id=batch_id
+    )
+
+    url_mappings = insert_urls_info.url_mappings
+    assert len(url_mappings) == 2
+    assert url_mappings[0].url == "https://example.com/1"
+    assert url_mappings[1].url == "https://example.com/2"
+
+
+    assert insert_urls_info.original_count == 2
+    assert insert_urls_info.duplicate_count == 1
