+++ tests/automated/integration/db/test_database_structure.py
