+++ tests/automated/integration/tasks/test_agency_preannotation_task.py
@@ -10,7 +10,8 @@ from src.collectors.source_collectors.muckrock.api_interface.lookup_response imp
 from src.collectors.source_collectors.muckrock.enums import AgencyLookupResponseType
 from src.core.tasks.operators.agency_identification.core import AgencyIdentificationTaskOperator
 from src.pdap_api.enums import MatchAgencyResponseStatus
-from tests.helpers.test_batch_creation_parameters import TestBatchCreationParameters, TestURLCreationParameters
+from tests.helpers.batch_creation_parameters.url_creation_parameters import TestURLCreationParameters
+from tests.helpers.batch_creation_parameters.core import TestBatchCreationParameters
 from src.db.models.core import Agency, AutomatedUrlAgencySuggestion
 from src.collectors.enums import CollectorType, URLStatus
 from src.core.tasks.enums import TaskOperatorOutcome
